allow to prepare the VMWare image
This is more of a showcase of a feature which I really like with NixOS.

If the CI or another machine has Nix installed, it's possible to
generate a new VMDK directly from the NixOS configuration. Building and
copying the VMDK would replace almost all the bootstrap steps from the
video. A similar technique is also possible to generate AMIs and other
system images from the same config.

I wasn't able to test this unfortunately because I don't have VMWare
installed. You might have to set a password to the user using the NixOS
config as well in order to be able to login.